+++ Matriz/pca.cpp
@@ -24,7 +24,7 @@ int main(int argc, char* argv[]){
 	vector<int> digitoRepr2 = imagenesTest.obtenerDigitos();
 
 	//Calculamos el cambio de base mediante pca
-	vector< vector<float> > cambioBase = imagenesTrain.pca(cantAutov,cantIterMetPot);
+	vector< vector<double> > cambioBase = imagenesTrain.pca(cantAutov,cantIterMetPot);
 	
 	//Le restamos la media del train y dividimos por sqrt(n-1) a las imagenes del test
 	imagenesTest.restarMedia(imagenesTrain);		
@@ -45,7 +45,7 @@ int main(int argc, char* argv[]){
 	//Hacemos el reconocimiento de digitos mediante kNN y comparamos los resultados con los valores reales
 	int aciertos = 0;
 	for(int i = 0; i<imagenesTest.Filas(); i++){
-		vector<float> fila = imagenesTestReducida.obtenerFila(i);
+		vector<double> fila = imagenesTestReducida.obtenerFila(i);
 		int res = imagenesTrainReducida.caenene(cantVecinos, fila);
 		if (res == digitoRepr2[i]){
 			aciertos++;
@@ -54,7 +54,7 @@ int main(int argc, char* argv[]){
 			cout << i << ": Funciona mal" << endl;
 		}
 	}
-	float hitRate = (float )aciertos/(float )imagenesTest.Filas();
+	double hitRate = (double )aciertos/(double )imagenesTest.Filas();
 	cout << hitRate << endl;
 
 	return 0;
